Refactoring chart property generator components
