+++ components/dashboards-web-component/src/gadgets-generation-wizard/components/chartPropertyGenerators/main/PieChart.jsx
