+++ components/dashboards-web-component/src/gadgets-generation-wizard/components/chartPropertyGenerators/sub/LineAreaBarChart.jsx
