+++ components/dashboards-web-component/src/gadgets-generation-wizard/components/chartPropertyGenerators/sub/ScatterChart.jsx
